move Archipelago.RetroArchClient.csproj into a subfolder to allow multiple projects
