+++ Archipelago.RetroArchClient.sln
@@ -1,6 +1,6 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Archipelago.RetroArchClient", "Archipelago.RetroArchClient.csproj", "{228A676D-272E-4665-A821-860F7F0D22FC}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Archipelago.RetroArchClient", "Archipelago.RetroArchClient/Archipelago.RetroArchClient.csproj", "{228A676D-272E-4665-A821-860F7F0D22FC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
