+++ Archipelago.RetroArchClient/OcarinaOfTime/Enums/Area.cs
@@ -1,0 +1,57 @@
+namespace Archipelago.RetroArchClient.OcarinaOfTime.Enums;
+
+// Maybe add documentation detailing what this enum is for, what each element represents, etc.
+// Helps to get other developers interested in helping with the client up to speed.
+
+// See Enums.GameModes for example on how this could be achieved.
+public enum Area
+{
+	// Overworld
+	KokiriForest,
+	LostWoods,
+	SacredForestMeadow,
+	HyruleField,
+	LonLonRanch,
+	Market,
+	HyruleCastle,
+	KakarikoVillage,
+	Graveyard,
+	DeathMountainTrail,
+	GoronCity,
+	DeathMountainCrater,
+	ZorasRiver,
+	ZorasDomain,
+	ZorasFountain,
+	LakeHylia,
+	GerudoValley,
+	GerudoFortress,
+	HauntedWasteland,
+	DesertColossus,
+	OutsideGanonsCastle,
+
+	// Dungeons
+	DekuTree,
+	DekuTreeMq,
+	DodongosCavern,
+	DodongosCavernMq,
+	JabuJabusBelly,
+	JabuJabusBellyMq,
+	ForestTemple,
+	ForestTempleMq,
+	FireTemple,
+	FireTempleMq,
+	IceCavern,
+	IceCavernMq,
+	WaterTemple,
+	WaterTempleMq,
+	BottomOfTheWell,
+	BottomOfTheWellMq,
+	ShadowTemple,
+	ShadowTempleMq,
+	GerudoTrainingGround,
+	GerudoTrainingGroundMq,
+	SpiritTemple,
+	SpiritTempleMq,
+	GanonsCastle,
+	GanonsCastleMq
+}
