+++ Archipelago.RetroArchClient/OcarinaOfTime/Enums/LocationType.cs
@@ -1,0 +1,31 @@
+namespace Archipelago.RetroArchClient.OcarinaOfTime.Enums;
+
+// Maybe add documentation detailing what this enum is for, what each element represents, etc.
+// Helps to get other developers interested in helping with the client up to speed.
+
+// See Enums.GameModes for example on how this could be achieved.
+public enum LocationType
+{
+	Chest,
+	Cow,
+	Skulltula,
+	Shop,
+	GroundItem,
+	Event,
+	GetInfo,
+	InfoTable,
+	Scrubsanity,
+	BossItem,
+	BigPoeBottle,
+	GreatFairy,
+	TrailGreatFairy,
+	CraterGreatFairy,
+	Medigoron,
+	BiggoronSword,
+	BeanSale,
+	FishingChild,
+	FishingAdult,
+	FireArrows,
+	MembershipCardCheck,
+	BombchuSalesman,
+}
