+++ Archipelago.RetroArchClient/Utils/ByteUtils.cs
@@ -1,0 +1,7 @@
+namespace Archipelago.RetroArchClient.Utils;
+
+public static class ByteUtils
+{
+	public static bool CheckBit(long memoryToCheck, byte bitToCheck)
+		=> (memoryToCheck & (1 << bitToCheck)) >= 1;
+}
