+++ BuildScripts/Win/publish-win-Release.bat
@@ -1,2 +1,2 @@
 @echo off
-dotnet publish ../../Archipelago.RetroArchClient.csproj --configuration Release --self-contained -r win-x64
+dotnet publish ../../Archipelago.RetroArchClient/Archipelago.RetroArchClient.csproj --configuration Release --self-contained -r win-x64
